Clean up issueCloned to issuer.js.
